requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
torch==2.0.1
numpy==1.25.0
Pillow==10.0.0
matplotlib==3.7.1
torchvision==0.15.2